replaced mariner mysql with mongodb
