+++ config/keys.js
@@ -1,0 +1,6 @@
+module.exports = {
+  mongoDb: {
+    dbUri: 'mongodb://alex:alexUser@ds019633.mlab.com:19633/comment-gets'
+  }
+}
+
