+++ micro-services/comment-retrieval/app.js
@@ -4,11 +4,10 @@ var favicon = require('serve-favicon');
 var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
-
 var index = require('./routes/index');
 var users = require('./routes/users');
 var test = require('./routes/test');
-
+var mongoDb = require('../database/')
 var app = express();
 
 // view engine setup
