+++ micro-services/comment-retrieval/package.json
@@ -13,6 +13,7 @@
     "ejs": "~2.5.7",
     "express": "~4.15.5",
     "morgan": "~1.9.0",
+    "mysql": "^2.15.0",
     "serve-favicon": "~2.4.5"
   }
 }
