+++ package.json
@@ -22,6 +22,7 @@
     "cookie-parser": "^1.4.3",
     "ejs": "^2.5.7",
     "express": "^4.16.2",
+    "mongoose": "^5.0.7",
     "morgan": "^1.9.0",
     "mysql": "^2.15.0",
     "sequelize": "^4.33.4",
